refactored donation form  and wired donate logic
