+++ src/components/Donator/donatorSchema.ts
@@ -5,4 +5,5 @@ export const donatorSchema = Yup.object().shape({
     .typeError("Amount must be a number")
     .required("Kindly specify amount")
     .positive("Amount must be greater than zero "),
+  split: Yup.number(),
 });
