+++ src/pages/TCA/TCA.tsx
@@ -1,6 +1,7 @@
 import Donator from "components/Donator/Donator";
+import TCAForm from "components/Donator/TCAForm";
 import { useGetToken } from "contexts/AuthProvider";
-import { Redirect, useRouteMatch } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 import { app, site } from "types/routes";
 
 export default function TCA() {
@@ -13,7 +14,9 @@ export default function TCA() {
 
   return (
     <div className="grid place-items-center min-h-screen">
-      <Donator />
+      <Donator>
+        <TCAForm />
+      </Donator>
     </div>
   );
 }
